+++ gfc-client/internal/upgrade/install_cmd_stub.go
@@ -1,0 +1,7 @@
+//go:build !unix
+
+package upgrade
+
+import "os/exec"
+
+func configureInstallCmd(cmd *exec.Cmd) {}